ensure that osu benchmarks have stopping condition
currently, the second worker can exit too early, and the first
is left waiting for it to come online. I tested if the batch job
would cut after one hour (the runtime limit) but the single VM
seemed to keep going. This is immensely dangerous (from a cost
standpoint) but a simple thing we can do is create a done marker,
and then only exit the second VM when it is present, and it will
be generated after the mpirun commands complete. This will close #1

Signed-off-by: vsoch <vsoch@users.noreply.github.com>